@@ -1,7 +1,5 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using Microsoft.Extensions.Configuration.AzureAppConfiguration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
@@ -9,17 +7,21 @@ namespace RobbyBot
 {
     public class Program
     {
-        public static void Main(string[] args)
-        {
+        public static void Main(string[] args) => 
             CreateHostBuilder(args).Build().Run();
-        }
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
+        static IHostBuilder CreateHostBuilder(string[] args) =>
             Host.CreateDefaultBuilder(args)
-                .ConfigureServices(services =>
+                .ConfigureAppConfiguration((hostingContext, config) =>
                 {
-                    services.AddHostedService<GameRequestHandler>();
-                    services.AddHostedService<ShapeHandler>();
-                });
+                    config.AddAzureAppConfiguration(options => 
+                    {
+                        var cnStr = Environment.GetEnvironmentVariable("AzureAppConfigConnectionString");
+                        options.Connect(cnStr);
+                    });
+                })
+                .ConfigureServices(services => 
+                    services.AddHostedService<GameRequestHandler>()
+                            .AddHostedService<ShapeHandler>());
     }
 }
